pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tko"
version = "2.6.6"
description = "Add your description here"
authors = [
    { name = "David", email = "sena.ufc@gmail.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "appdirs>=1.4.4",
    "chardet>=5.2.0",
    "diff-match-patch>=20241021",
    "pydot>=3.0.3",
    "pytest>=8.3.4",
    "pyyaml>=6.0.2",
    "uniplot>=0.16.2",
    "windows-curses; platform_system=='Windows'"
]

[project.scripts]
tko = "tko.__main__:main"
